backend/requirements.txt
asgiref==3.6.0
cachetools==5.3.1
captcha==0.5.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.7
django-allauth==0.54.0
django-cors-headers==3.14.0
django-crispy-forms==2.1
django-crontab==0.7.1
django-phonenumber-field==7.2.0
django-recaptcha==4.0.0
django-rest-auth==0.9.5
django-tinymce==3.6.1
djangorestframework==3.14.0
filelock==3.13.1
fontawesomefree==6.5.1
future==0.18.3
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
phonenumbers==8.13.26
Pillow==9.5.0
platformdirs==4.1.0
protobuf==4.23.4
psycopg2==2.9.8
psycopg2-binary==2.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.25.0
whitenoise==6.5.0
